requirements.txt
fastapi
uvicorn[standard]
openai
pydantic
pyyaml
python-dotenv
pytest
pytest-asyncio
slowapi
limits
redis
prometheus-client